# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "2.3.0",
    "title": "Task Pro API",
    "description": "Task Pro API provides endpoints for managing projects, tasks, and user assignments with secure authentication and efficient workflows.",
    "contact": {
      "name": "TaskPro",
      "url": "https://www.taskpro.qzz.io"
    }
  },
  "servers": [
    { "url": "http://localhost:9537/api" },
    { "url": "https://api.taskpro.qzz.io" }
  ],
  "paths": {
...
    "/user/settings": {
      "get": {
        "tags": ["User"],
        "operationId": "getAllSettings",
        "summary": "Get all user settings",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/GetAllSettingsResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/user/settings/general": {
      "patch": {
        "tags": ["User"],
        "operationId": "updateGeneralSettings",
        "summary": "Update general settings",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "theme": {
                    "type": "string",
                    "enum": ["light", "dark", "system"],
                    "example": "light"
                  },
                  "accentColor": {
                    "type": "string",
                    "enum": [
                      "blue",
                      "purple",
                      "red",
                      "gray",
                      "cyan",
                      "yellow",
                      "indigo",
                      "green"
                    ],
                    "example": "blue"
                  },
                  "firstDayOfWeek": {
                    "type": "string",
                    "enum": ["monday", "sunday"],
                    "example": "monday"
                  },
                  "dateFormat": {
                    "type": "string",
                    "enum": ["dd_mm_yyyy", "mm_dd_yyyy", "yyyy_mm_dd"],
                    "example": "dd_mm_yyyy"
                  },
                  "boardBackgroundBlur": {
                    "type": "string",
                    "enum": ["off", "low", "medium"],
                    "example": "off"
                  },
                  "usePointerCursors": {
                    "type": "boolean",
                    "example": true
                  },
                  "enableAnimations": {
                    "type": "boolean",
                    "example": true
                  },
                  "confirmBeforeDelete": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GeneralSettingsResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/user/settings/task": {
      "patch": {
        "tags": ["User"],
        "operationId": "updateTaskSettings",
        "summary": "Update task settings",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sortTasksBy": {
                    "type": "string",
                    "enum": [
                      "manual",
                      "priority",
                      "deadline",
                      "created",
                      "alphabetical"
                    ],
                    "example": "manual"
                  },
                  "defaultPriority": {
                    "type": "string",
                    "$ref": "#/components/schemas/TaskPriority",
                    "example": "without"
                  },
                  "defaultDeadline": {
                    "type": "string",
                    "enum": [
                      "none",
                      "today",
                      "tomorrow",
                      "three_days",
                      "one_week"
                    ],
                    "example": "none"
                  },
                  "cardDensity": {
                    "type": "string",
                    "enum": ["compact", "comfortable"],
                    "example": "compact"
                  },
                  "showCompletedTasks": {
                    "type": "boolean",
                    "example": true
                  },
                  "showPriorityIndicator": {
                    "type": "boolean",
                    "example": true
                  },
                  "newTaskPosition": {
                    "type": "string",
                    "enum": ["top", "bottom"],
                    "example": "bottom"
                  },
                  "enableNaturalLanguageDates": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/TaskSettingsResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/user/settings/label": {
      "patch": {
        "tags": ["User"],
        "operationId": "updateLabelSettings",
        "summary": "Update label settings",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "showLabelsOnTask": {
                    "type": "boolean",
                    "example": true
                  },
                  "labelDisplay": {
                    "type": "string",
                    "enum": ["full", "compact"],
                    "example": "full"
                  },
                  "maxLabelsShown": {
                    "type": "string",
                    "enum": ["one", "two", "three", "all"],
                    "example": "three"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/LabelSettingsResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/user/settings/accessibility": {
      "patch": {
        "tags": ["User"],
        "operationId": "updateAccessibilitySettings",
        "summary": "Update accessibility settings",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessibilitySettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AccessibilitySettingsResponse"
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/task/{columnId}": {
      "post": {
        "tags": ["Task"],
        "operationId": "createTask",
        "summary": "Create new task",
        "parameters": [
          {
            "in": "path",
            "name": "columnId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "priority"],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Publication of the project"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Review the project materials: Familiarize yourself with the project's content, including text, images, and any supplementary materials."
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-10-15"
                  },
                  "priority": { "$ref": "#/components/schemas/TaskPriority" },
                  "labels": {
                    "type": "array",
                    "description": "Array of Label ids to attach to the task",
                    "items": { "type": "string" },
                    "example": ["6672fdccc07147fc7ae1bb93"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/TaskResponse" },
          "400": { "$ref": "#/components/responses/TaskBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/task/{columnId}/order": {
      "patch": {
        "tags": ["Task"],
        "operationId": "updateTasksOrder",
        "summary": "Update tasks order",
        "parameters": [
          {
            "in": "path",
            "name": "columnId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ids"],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "6672fdccc07147fc7ae1bb93",
                      "6743474be4006eca6c8122e7"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/UpdateTaskOrderResponse" },
          "400": { "$ref": "#/components/responses/UpdateOrderBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/task/{taskId}": {
      "patch": {
        "tags": ["Task"],
        "operationId": "updateTask",
        "summary": "Update task by id",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Final Project Review"
                  },
                  "description": {
                    "type": ["string", "null"],
                    "minLength": 3,
                    "example": "Review and finalize all project documentation."
                  },
                  "deadline": {
                    "type": ["string", "null"],
                    "format": "date-time",
                    "example": "2025-11-01"
                  },
                  "priority": {
                    "allOf": [{ "$ref": "#/components/schemas/TaskPriority" }],
                    "example": "low"
                  },
                  "completed": {
                    "type": "boolean",
                    "example": true
                  },
                  "columnId": {
                    "type": "string",
                    "example": "6672fdccc07147fc7ae1bb93",
                    "description": "Id of the column to move the task to"
                  },
                  "labels": {
                    "type": "array",
                    "description": "Array of Label ids to attach to the task",
                    "items": { "type": "string" },
                    "example": ["6672fdccc07147fc7ae1bb93"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/TaskResponse" },
          "400": { "$ref": "#/components/responses/TaskBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Task"],
        "operationId": "deleteTask",
        "summary": "Delete task by id",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "The resource was deleted successfully." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/label": {
      "get": {
        "tags": ["Label"],
        "operationId": "getAllLabels",
        "summary": "Get all labels",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/AllLabelsResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Label"],
        "operationId": "createLabel",
        "summary": "Create new label",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "color"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "Bug"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Description of the bug"
                  },
                  "color": { "$ref": "#/components/schemas/AccentColor" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/LabelResponse" },
          "400": { "$ref": "#/components/responses/LabelBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "priority",
          "deadline",
          "completed",
          "columnId",
          "order",
          "completedAt",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672fdccc07147fc7ae1bb93"
          },
          "title": {
            "type": "string",
            "example": "Publication of the project"
          },
          "description": {
            "type": ["string", "null"],
            "example": "Review the project materials: Familiarize yourself with the project's content, including text, images, and any supplementary materials."
          },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "deadline": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": "2025-10-15T00:00:00.000Z"
          },
          "order": { "type": "integer", "example": 1 },
          "completed": { "type": "boolean", "example": false },
          "completedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": null
          },
          "columnId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          },
          "labels": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Label" }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          }
        }
      },
      "TaskPriority": {
        "type": "string",
        "enum": ["without", "low", "medium", "high"],
        "example": "high"
      },
      "Label": {
        "type": "object",
        "required": ["id", "name", "color", "userId", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672fdccc07147fc7ae1bb93"
          },
          "name": {
            "type": "string",
            "example": "Bug"
          },
          "description": {
            "type": ["string", "null"],
            "example": "Description of the bug"
          },
          "color": { "$ref": "#/components/schemas/AccentColor" },
          "userId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          },
          "taskId": {
            "type": ["string", "null"],
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          }
        }
      },
      "AccentColor": {
        "type": "string",
        "enum": [
          "blue",
          "purple",
          "red",
          "gray",
          "cyan",
          "yellow",
          "indigo",
          "green"
        ],
        "example": "red"
      },
      "GeneralSettings": {
        "type": "object",
        "required": [
          "id",
          "theme",
          "accentColor",
          "firstDayOfWeek",
          "boardBackgroundBlur",
          "usePointerCursors",
          "enableAnimations",
          "confirmBeforeDelete",
          "userId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672fdccc07147fc7ae1bb93"
          },
          "theme": {
            "type": "string",
            "enum": ["light", "dark", "system"],
            "default": "light"
          },
          "accentColor": {
            "type": "string",
            "enum": [
              "blue",
              "purple",
              "red",
              "gray",
              "cyan",
              "yellow",
              "indigo",
              "green"
            ],
            "default": "blue"
          },
          "firstDayOfWeek": {
            "type": "string",
            "enum": ["monday", "sunday"],
            "example": "monday"
          },
          "dateFormat": {
            "type": "string",
            "enum": ["dd_mm_yyyy", "mm_dd_yyyy", "yyyy_mm_dd"],
            "example": "dd_mm_yyyy"
          },
          "boardBackgroundBlur": {
            "type": "string",
            "enum": ["off", "low", "medium"],
            "example": "off"
          },
          "usePointerCursors": {
            "type": "boolean",
            "example": true
          },
          "enableAnimations": {
            "type": "boolean",
            "example": true
          },
          "confirmBeforeDelete": {
            "type": "boolean",
            "example": true
          },
          "userId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          }
        }
      },
      "TaskSettings": {
        "type": "object",
        "required": [
          "id",
          "sortTasksBy",
          "defaultPriority",
          "defaultDeadline",
          "cardDensity",
          "showCompletedTasks",
          "showPriorityIndicator",
          "newTaskPosition",
          "enableNaturalLanguageDates",
          "userId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672fdccc07147fc7ae1bb93"
          },
          "sortTasksBy": {
            "type": "string",
            "enum": [
              "manual",
              "priority",
              "deadline",
              "created",
              "alphabetical"
            ],
            "example": "manual"
          },
          "defaultPriority": {
            "type": "string",
            "$ref": "#/components/schemas/TaskPriority",
            "example": "without"
          },
          "defaultDeadline": {
            "type": "string",
            "enum": ["none", "today", "tomorrow", "three_days", "one_week"],
            "example": "none"
          },
          "cardDensity": {
            "type": "string",
            "enum": ["compact", "comfortable"],
            "example": "compact"
          },
          "showCompletedTasks": {
            "type": "boolean",
            "example": true
          },
          "showPriorityIndicator": {
            "type": "boolean",
            "example": true
          },
          "newTaskPosition": {
            "type": "string",
            "enum": ["top", "bottom"],
            "example": "bottom"
          },
          "enableNaturalLanguageDates": {
            "type": "boolean",
            "example": true
          },
          "userId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          }
        }
      },
      "LabelSettings": {
        "type": "object",
        "required": [
          "id",
          "showLabelsOnTask",
          "labelDisplay",
          "maxLabelsShown",
          "userId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672fdccc07147fc7ae1bb93"
          },
          "showLabelsOnTask": {
            "type": "boolean",
            "example": true
          },
          "labelDisplay": {
            "type": "string",
            "enum": ["full", "compact"],
            "example": "full"
          },
          "maxLabelsShown": {
            "type": "string",
            "enum": ["one", "two", "three", "all"],
            "example": "three"
          },
          "userId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          }
        }
      },
      "AccessibilitySettings": {
        "type": "object",
        "required": [
          "id",
          "fontSize",
          "reducedMotion",
          "highContrast",
          "focusIndicators",
          "keyboardNavigationHints",
          "userId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672fdccc07147fc7ae1bb93"
          },
          "fontSize": {
            "type": "string",
            "enum": ["small", "medium", "large"],
            "exmaple": "medium"
          },
          "reducedMotion": {
            "type": "boolean",
            "exmaple": false
          },
          "highContrast": {
            "type": "boolean",
            "exmaple": false
          },
          "focusIndicators": {
            "type": "boolean",
            "exmaple": true
          },
          "keyboardNavigationHints": {
            "type": "boolean",
            "exmaple": false
          },
          "userId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer", "example": 400 },
          "message": {
            "oneOf": [{ "type": "string" }, { "type": "object" }]
          }
        }
      }
    },
    "responses": {
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": { "statusCode": 404, "message": "Not found" }
            }
          }
        }
      },
...
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": { "statusCode": 401, "message": "Unauthorized" }
            }
          }
        }
      },
...
      "UpdateOrderBadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": {
                "statusCode": 400,
                "message": {
                  "ids": "The field must be of type array"
                }
              }
            }
          }
        }
      },
      "GetAllSettingsResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "allOf": [
                {
                  "type": "object",
                  "required": ["general", "task", "label", "accessibility"],
                  "properties": {
                    "general": {
                      "$ref": "#/components/schemas/GeneralSettings"
                    },
                    "task": { "$ref": "#/components/schemas/TaskSettings" },
                    "label": { "$ref": "#/components/schemas/LabelSettings" },
                    "accessibility": {
                      "$ref": "#/components/schemas/AccessibilitySettings"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "GeneralSettingsResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/GeneralSettings"
            }
          }
        }
      },
      "TaskSettingsResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/TaskSettings"
            }
          }
        }
      },
      "LabelSettingsResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/LabelSettings"
            }
          }
        }
      },
      "AccessibilitySettingsResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/AccessibilitySettings"
            }
          }
        }
      },
...
      "TaskResponse": {
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Task" }
          }
        }
      },
      "TaskBadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": {
                "statusCode": 400,
                "message": {
                  "description": "The field must be of type string",
                  "priority": "The field must be one of the following values: without, low, medium, high",
                  "columnId": "The field must have a valid ObjectId format"
                }
              }
            }
          }
        }
      },
      "UpdateTaskOrderResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": { "$ref": "#/components/schemas/Task" }
            }
          }
        }
      },
      "AllLabelsResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": { "$ref": "#/components/schemas/Label" }
            }
          }
        }
      },
      "LabelResponse": {
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Label" }
          }
        }
      },
      "LabelBadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": {
                "statusCode": 400,
                "message": {
                  "color": "The field must be one of the following values: blue, purple, red, gray, cyan, yellow, indigo, green"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}